use a concern for archived users in controller, replace banner with flash content
