+++ app/controllers/users/follow_ups_controller.rb
@@ -5,6 +5,7 @@ module Users
                   :set_user_is_archived, :set_back_to_users_list_url, only: [:index]
 
     include BackToListConcern
+    include UserArchivedConcern
     include Users::Taggable
     include Users::Archivable
 
@@ -62,16 +63,6 @@ module Users
 
     def set_user_archives
       @user_archives = @user.archives
-    end
-
-    def set_user_is_archived
-      @user_is_archived =
-        @user.archives.where(organisation: user_agent_department_organisations).count ==
-        user_agent_department_organisations.count
-    end
-
-    def user_agent_department_organisations
-      @user_agent_department_organisations ||= @user.organisations & @current_organisations
     end
 
     def set_current_organisations
