+++ app/controllers/users/parcours_controller.rb
@@ -4,6 +4,7 @@ module Users
                   :set_current_organisations, :set_user_archives, :set_user_is_archived, only: [:show]
 
     include BackToListConcern
+    include UserArchivedConcern
     include Users::Taggable
     include Users::Archivable
 
@@ -28,16 +29,6 @@ module Users
 
     def set_user_archives
       @user_archives = @user.archives
-    end
-
-    def set_user_is_archived
-      @user_is_archived =
-        @user.archives.where(organisation: user_agent_department_organisations).count ==
-        user_agent_department_organisations.count
-    end
-
-    def user_agent_department_organisations
-      @user_agent_department_organisations ||= @user.organisations & @current_organisations
     end
 
     def set_current_organisations
