+++ app/javascript/stylesheets/components/_banner.scss
@@ -11,15 +11,3 @@
     color: white;
   }
 }
-
-.archive-banner {
-  background-color: $orange-light;
-  color: $dark-blue;
-  padding: 13px 0;
-  position: sticky;
-  top: 0;
-  p {
-    font-size: 1.1rem;
-    margin: 0;
-  }
-}
