--- app/views/users/_archive_banner.html.erb
@@ -1,5 +1,0 @@
-<div id="archive-banner" class="fixed-top archive-banner">
-  <p class="text-center">
-    Dossier archivé
-  </p>
-</div>
