+++ app/views/users/show.html.erb
@@ -1,7 +1,7 @@
 <%= turbo_frame_tag "user_page", data: { turbo_action: "advance" } do %>
   <%= render "users/page_title", user: @user %>
 
-  <%= render 'archive_banner' if @user_is_archived %>
+  <%= render "common/flash_content", type: "notice", message: @archived_message if @user_is_archived %>
 
   <div class="container text-dark-blue h4-as-labels mt-4">
     <%= render "user_page_header" %>
